backend/requirements.txt
aiofiles
asgiref
bcrypt
cffi
click
colorama
ecdsa
fastapi
greenlet
h11
Jinja2
MarkupSafe
passlib
pyasn1
pycparser
pydantic
python-decouple
python-jose
python-multipart
rsa
six
SQLAlchemy
starlette
typing-extensions
uvicorn
pre-commit
pydantic_settings
mysqlclient
pytz
pymysql
pydantic[email]
requests
loguru